renamed reducers, actions, added asunc device add
