--- src/actions/devicesActions.js
@@ -1,9 +1,0 @@
-import { ADD_DEVICE } from '../constants/constants';
-
-export const addDevice = (device) => {
-  return {
-    type: ADD_DEVICE,
-    device
-  };
-};
-
