+++ src/reducers/addDevice.reducer.js
