+++ src/reducers/reducer.js
@@ -1,7 +1,7 @@
 import searchAndFilter from './searchAndFilter.reducer';
 import { combineReducers } from 'redux';
-import deviceProto from './deviceProtoReducer';
-import devices from './devicesReducer';
+import deviceProto from './builder.reducer';
+import devices from './addDevice.reducer';
 import devicesList from './devicesList.reducer';
 
 const rootReducer = combineReducers({
